rag-server/pyproject.toml
[project]
name = "rag-server"
version = "0.1.0"
description = "AI/RAG pipeline for Story Arc routing — ChromaDB + Gemini gatekeeper"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "chromadb>=1.0.0",
    "langchain-google-genai>=2.1.0",
    "langchain-core>=0.3.0",
    "pydantic>=2.0.0",
    "pymongo>=4.8.0",
    "python-dotenv>=1.1.0",
]

[project.scripts]
dev = "uvicorn main:app --reload --port 8100"
